# doc/api.yaml
openapi: 3.0.0
paths:
...
  /artist/{id}:
    get:
      description: This endpoint retrieves an artist by their ID.
      operationId: ArtistsController_getArtist
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the artist (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174001
            type: string
      responses:
        "200":
          description: The artist has been successfully retrieved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "404":
          description: Artist not found. The artist with the specified ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Artist with ID 123e4567-e89b-12d3-a456-426614174001 not found
                  error: Not Found
      security: *a3
      summary: Get an artist by ID
      tags: *a4
    put:
      description: This endpoint updates an artist by their ID with new data.
      operationId: ArtistsController_updateArtist
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the artist (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174001
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateArtistDto"
      responses:
        "200":
          description: The The user password has been successfully updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "404":
          description: Artist not found. The artist with the specified ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Artist with ID 123e4567-e89b-12d3-a456-426614174001 not found
                  error: Not Found
      security: *a3
      summary: Update an artist by ID
      tags: *a4
    delete:
      description: This endpoint deletes an artist by their ID.
      operationId: ArtistsController_deleteArtist
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the artist (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174001
            type: string
      responses:
        "204":
          description: The artist has been successfully deleted.
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "404":
          description: Artist not found. The artist with the specified ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Artist with ID 123e4567-e89b-12d3-a456-426614174001 not found
                  error: Not Found
      security: *a3
      summary: Delete an artist by ID
      tags: *a4
...
  /album/{id}:
    get:
      description: This endpoint retrieves an album by its ID.
      operationId: AlbumsController_getAlbum
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the album (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174002
            type: string
      responses:
        "200":
          description: The album has been successfully retrieved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "404":
          description: Album not found. The album with the specified ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Album with ID 123e4567-e89b-12d3-a456-426614174002 not found
                  error: Not Found
      security: *a5
      summary: Get an album by ID
      tags: *a6
    put:
      description: This endpoint updates an existing album by its ID with new data.
      operationId: AlbumsController_updateAlbum
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the album (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174002
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlbumDto"
      responses:
        "200":
          description: The The user password has been successfully updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "404":
          description: Album not found. The album with the specified ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Album with ID 123e4567-e89b-12d3-a456-426614174002 not found
                  error: Not Found
        "422":
          description: Artist not exist. The artist with the specified ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 422
                  message: Artist with ID 123e4567-e89b-12d3-a456-426614174001 does not exist
                  error: Not Found
      security: *a5
      summary: Update an album by ID
      tags: *a6
    delete:
      description: This endpoint deletes an album by its ID.
      operationId: AlbumsController_deleteAlbum
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the album (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174002
            type: string
      responses:
        "204":
          description: The album has been successfully deleted.
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "404":
          description: Album not found. The album with the specified ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Album with ID 123e4567-e89b-12d3-a456-426614174002 not found
                  error: Not Found
      security: *a5
      summary: Delete an album by ID
      tags: *a6
...
  /track/{id}:
    get:
      description: This endpoint retrieves an track by its ID.
      operationId: TracksController_getTrack
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the track (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174003
            type: string
      responses:
        "200":
          description: The track has been successfully retrieved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "404":
          description: Track not found. The track with the specified ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Track with ID 123e4567-e89b-12d3-a456-426614174003 not found
                  error: Not Found
      security: *a7
      summary: Get an track by ID
      tags: *a8
    put:
      description: This endpoint updates an existing track by its ID with new data.
      operationId: TracksController_updateTrack
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the track (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174003
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrackDto"
      responses:
        "200":
          description: The The user password has been successfully updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "404":
          description: Track not found. The track with the specified ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Track with ID 123e4567-e89b-12d3-a456-426614174003 not found
                  error: Not Found
        "422":
          description: Artist or album not exist. The artist or album with the specified
            ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 422
                  message: Artist with ID 123e4567-e89b-12d3-a456-426614174001 does not exist
                  error: Not Found
      security: *a7
      summary: Update an track by ID
      tags: *a8
    delete:
      description: This endpoint deletes an track by its ID.
      operationId: TracksController_deleteTrack
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the track (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174003
            type: string
      responses:
        "204":
          description: The track has been successfully deleted.
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "404":
          description: Track not found. The track with the specified ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Track with ID 123e4567-e89b-12d3-a456-426614174003 not found
                  error: Not Found
      security: *a7
      summary: Delete an track by ID
      tags: *a8
  /favs:
    get:
      description: This endpoint retrieves all the favorites.
      operationId: FavoritesController_getAllFavorites
      parameters: []
      responses:
        "200":
          description: The The favorites has been successfully retrieved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Favorites"
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
      security: &a9
        - access-token: []
      summary: Get all favorites
      tags: &a10
        - Favorites
  /favs/artist/{id}:
    post:
      description: This endpoint adds a artist to the favorites.
      operationId: FavoritesController_addArtistToFavorites
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the artist (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174001
            type: string
      responses:
        "201":
          description: The artist has been successfully added to favorites.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddedFavorite"
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "422":
          description: Artist not exist. The artist with the specified ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 422
                  message: Artist with ID 123e4567-e89b-12d3-a456-426614174001 does not exist
                  error: Not Found
      security: *a9
      summary: Add a artist to favorites
      tags: *a10
    delete:
      description: This endpoint removes a artist from the favorites.
      operationId: FavoritesController_removeArtistFromFavorites
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the artist (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174001
            type: string
      responses:
        "204":
          description: The artist has been successfully removed from favorites.
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "404":
          description: Artist not found in favorites. The artist with the specified ID
            does not exist in favorites.
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Artist with ID 123e4567-e89b-12d3-a456-426614174001 not found in
                    favorites
                  error: Not Found
      security: *a9
      summary: Remove a artist from favorites
      tags: *a10
  /favs/album/{id}:
    post:
      description: This endpoint adds a album to the favorites.
      operationId: FavoritesController_addAlbumToFavorites
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the album (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174002
            type: string
      responses:
        "201":
          description: The album has been successfully added to favorites.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddedFavorite"
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "422":
          description: Album not exist. The album with the specified ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 422
                  message: Album with ID 123e4567-e89b-12d3-a456-426614174002 does not exist
                  error: Not Found
      security: *a9
      summary: Add a album to favorites
      tags: *a10
    delete:
      description: This endpoint removes a album from the favorites.
      operationId: FavoritesController_removeAlbumFromFavorites
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the album (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174002
            type: string
      responses:
        "204":
          description: The album has been successfully removed from favorites.
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "404":
          description: Album not found in favorites. The album with the specified ID does
            not exist in favorites.
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Album with ID 123e4567-e89b-12d3-a456-426614174002 not found in
                    favorites
                  error: Not Found
      security: *a9
      summary: Remove a album from favorites
      tags: *a10
  /favs/track/{id}:
    post:
      description: This endpoint adds a track to the favorites.
      operationId: FavoritesController_addTrackToFavorites
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the track (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174003
            type: string
      responses:
        "201":
          description: The track has been successfully added to favorites.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddedFavorite"
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "422":
          description: Track not exist. The track with the specified ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 422
                  message: Track with ID 123e4567-e89b-12d3-a456-426614174003 does not exist
                  error: Not Found
      security: *a9
      summary: Add a track to favorites
      tags: *a10
    delete:
      description: This endpoint removes a track from the favorites.
      operationId: FavoritesController_removeTrackFromFavorites
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the track (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174003
            type: string
      responses:
        "204":
          description: The track has been successfully removed from favorites.
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "404":
          description: Track not found in favorites. The track with the specified ID does
            not exist in favorites.
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Track with ID 123e4567-e89b-12d3-a456-426614174003 not found in
                    favorites
                  error: Not Found
      security: *a9
      summary: Remove a track from favorites
      tags: *a10
...
info:
  title: Home Library Service
  description: API for managing music albums, artists, and tracks
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: Authorization
      description: Enter JWT access token
      in: header
  schemas:
...
    Artist:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174001
        name:
          type: string
          example: The Beatles
        grammy:
          type: boolean
          description: Specifies if the artist has a Grammy Award
          example: false
    UpdateArtistDto:
      type: object
      properties:
        name:
          type: string
          description: The name of the artist
          example: The Beatles
        grammy:
          type: boolean
          description: Specifies if the artist has a Grammy Award
          example: false
...
    Album:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174002
        name:
          type: string
          example: Abbey Road
        year:
          type: number
          description: The year the album was released
          example: 1969
        artistId:
          type: string
          description: The artist ID associated with the album
          nullable: true
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174001
    UpdateAlbumDto:
      type: object
      properties:
        name:
          type: string
          description: The name of the album
          example: Abbey Road
        year:
          type: number
          description: The year the album was released
          example: 1969
        artistId:
          type: string
          description: The artist ID associated with the album
          nullable: true
          format: uuid
...
    UpdateTrackDto:
      type: object
      properties:
        name:
          type: string
          description: The name of the track
          example: Come Together
        duration:
          type: number
          description: The duration of the track in seconds
          example: 201
        artistId:
          type: string
          description: The artist ID associated with the track
          nullable: true
          format: uuid
        albumId:
          type: string
          description: The album ID associated with the track
          nullable: true
          format: uuid
...